+++ Pengo/PengoCharacter.cpp
@@ -9,6 +9,7 @@
 #include "IdleState.h"
 #include "MovingState.h"
 #include "DeadState.h"
+#include "Layers.h"
 
 using namespace dae;
 
@@ -29,7 +30,7 @@ PengoCharacter::PengoCharacter(PengoColor pengoColor)
 	textureComponent->SetTexture("characterSprites.png", 
 								SDL_Rect{ GetTextureOffsetX(pengoColor), GetTextureOffsetY(pengoColor), 16, 16}, 1.5f);
 	auto boxColliderComponent = std::make_unique<BoxColliderComponent>(16 * 1.5f, 16 * 1.5f, ObjectType::movable, 
-																	   m_CharacterObject.get(), Layer::pengo);
+																	   m_CharacterObject.get(), static_cast<uint32_t>(Layer::pengo));
 
 	m_CharacterObject->AddComponent(std::move(healthComponent));
 	m_CharacterObject->AddComponent(std::move(pointComponent));
